pyproject.toml
[tool.poetry]
name = "thumbnailAPI"
version = "0.1.0"
description = "a simple api for resizing image to thumbnail"

authors = ["Tinghe Zhang <tzhang329@gmail.com>"]

license = "MIT"

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = {version = "0.18.3", extras = ["standard"]}
fastapi = "0.82.0"
prometheus-client = "^0.15.0"
prometheus-fastapi-instrumentator = "^5.9.1"
pydantic = "^1.10.2"
celery = "^5.2.7"
redis = "^4.3.4"
validators = "^0.20.0"
requests = "^2.28.1"
pymongo = "^4.3.2"
pillow = "^9.3.0"
pytest-celery = "^0.0.0"

[tool.poetry.dev-dependencies]
requests = "2.28.1"
pytest = "7.1.3"
pytest-cov = "3.0.0"
coverage = "6.4.4"
black = "22.8.0"
pre-commit = "2.17.0"

# pytest configs  https://github.com/pytest-dev/pytest-cov
# https://docs.pytest.org/en/latest/reference/customize.html
[tool.pytest.ini_options]
testpaths = [
    "tests",
]

# https://black.readthedocs.io/en/stable/usage_and_configuration/the_basics.html#configuration-via-a-file
[tool.black]
line-length = 80
target-version = ["py311"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
